Refactor loading out of MicrophysicsHook
The __init__ of MicrophysicsHook was doing too much and required frequent
editing.

Use dependency injection to decouple things and enable a better separation for
concerns.
