+++ external/emulation/emulation/__init__.py
@@ -1,3 +1,12 @@
+# flake8: noqa
+# Tensorflow looks at sys args which are not initialized
+# when this module is loaded under callpyfort, so ensure
+# it's available here
+import sys
+
+if not hasattr(sys, "argv"):
+    sys.argv = [""]
+
 from emulation.config import (
     get_hooks,
     EmulationConfig,
